+++ src/main/java/org/opensearch/security/ssl/OpenSearchSecuritySSLPlugin.java
@@ -18,6 +18,7 @@
 package org.opensearch.security.ssl;
 
 import java.nio.file.Path;
+import java.security.Security;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.Collections;
@@ -32,6 +33,7 @@ import java.util.function.Supplier;
 import com.fasterxml.jackson.databind.InjectableValues;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
+import org.bouncycastle.jcajce.provider.BouncyCastleFipsProvider;
 
 import org.opensearch.OpenSearchException;
 import org.opensearch.Version;
@@ -221,6 +223,8 @@ public class OpenSearchSecuritySSLPlugin extends Plugin implements SystemIndexPl
         if (!httpSSLEnabled && !transportSSLEnabled) {
             log.error("SSL not activated for http and/or transport.");
         }
+
+        tryAddSecurityProvider();
 
         this.sslSettingsManager = new SslSettingsManager(new Environment(settings, configPath));
     }
@@ -737,5 +741,15 @@ public class OpenSearchSecuritySSLPlugin extends Plugin implements SystemIndexPl
 
     public ThreadPool getThreadPool() {
         return this.threadPool;
+    }
+
+    private void tryAddSecurityProvider() {
+        AccessController.doPrivileged(() -> {
+            if (Security.getProvider("BCFIPS") == null) {
+                Security.addProvider(new BouncyCastleFipsProvider());
+                log.debug("Bouncy Castle FIPS Provider added");
+            }
+            return null;
+        });
     }
 }